+++ example/momotaro.py
@@ -12,7 +12,7 @@ from builder.world import World
 from builder.writer import Writer
 ## local files
 from assets import basic, accessory
-from config import DAYS, ITEMS, LAYERS, PERSONS, RUBIS, STAGES, TIMES, WORDS
+from config import AREAS, DAYS, ITEMS, LAYERS, PERSONS, RUBIS, STAGES, TIMES, WORDS
 
 
 W = Writer
@@ -61,6 +61,7 @@ def sc_get_peach(w: World):
             gma.have(w.peach, "そう考え、何とか木の棒を使ってこちらに引き寄せ、ひと抱えもある桃を手に入れました"),
             gma.go("それを籠に入れ、家に帰りました"),
             camera=w.granma,
+            area=w.Vila,
             stage=w.on_river,
             day=w.in_birth, time=w.at_midmorning,
             )
@@ -178,6 +179,7 @@ def sc_island(w: World):
             daemon.talk("参りました。どうか、命だけは……"),
             taro.talk("もうこれ以上村の人に悪さをするんじゃないぞ？", "いいな？"),
             taro.explain("こうして鬼たちから宝物を巻き上げ、村に持ち帰りました"),
+            area=w.DaemonIsland,
             stage=w.on_island_int,
             )
 
@@ -270,7 +272,7 @@ def create_world():
     w.setAssets(accessory.ASSET)
     #   set DB (user)
     w.buildDB(PERSONS,
-            STAGES, DAYS, TIMES, ITEMS, WORDS,
+            AREAS, STAGES, DAYS, TIMES, ITEMS, WORDS,
             RUBIS, LAYERS)
     #   set base year
     w.setBaseDate(113)