+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [0.4.4-4] - 2020-01-29
+### Added
+- Asset: acccessory
+### Changed
+- block output using list switch
 ### Fixed
 - time delta in Stageline